- Cleaned up TestOpacityIndex
